Deduplicate code between ag1 exporter and viewer
